requirements.txt
alembic==1.4.3
altgraph==0.17
aniso8601==8.0.0
click==7.1.2
et-xmlfile==1.0.1
Flask==1.1.2
Flask-Admin==1.5.6
Flask-HTTPAuth==4.1.0
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
fpdf2==2.0.5
future==0.18.2
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
numpy==1.19.2
openpyxl==3.0.5
pefile==2019.4.18
Pillow==6.2.2
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.8
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
pywin32-ctypes==0.2.0
six==1.15.0
SQLAlchemy==1.3.18
Werkzeug==1.0.1
WTForms==2.3.3
